Fix PredicateSet example in README.
This solves two issues with the `PredicateSet` example code. First, it removes the dependency on Darwin's `floor` by sending the number through `Int` and back to `Float` to get the integer value. Second, it fixes a bug in the declaration of `Q` — `{ $0 as Float }` doesn't return a `Bool`, as required by the set's initializer.